edit card updated

+++ react-trello/src/components/edit-card/index.js
@@ -12,6 +12,15 @@ import Transparent from '../transparent'
 import { useSocket } from '../../contexts/SocketProvider'
 import pic1 from '../../images/edit-card/pic1.svg'
 import pic2 from '../../images/edit-card/pic2.svg'
+import pic3 from '../../images/edit-card/pic3.svg'
+import pic4 from '../../images/edit-card/pic4.svg'
+import pic5 from '../../images/edit-card/pic5.svg'
+import pic6 from '../../images/edit-card/pic6.svg'
+import pic7 from '../../images/edit-card/pic7.svg'
+import pic8 from '../../images/edit-card/pic8.svg'
+import pic9 from '../../images/edit-card/pic9.svg'
+
+
 
 import Avatar from 'react-avatar'
 import ButtonClean from '../button-clean'
@@ -55,7 +64,12 @@ export default function EditCard(props) {
     }
 
     const deleteCard = useCallback(async (event) => {
+
         event.preventDefault()
+
+        if (!window.confirm('Are you sure you wish to delete this item?')) {
+            return
+        }
         const token = getCookie("x-auth-token")
         const response = await fetch(`http://localhost:4000/api/projects/lists/cards/${listId}/${cardId}`, {
             method: "DELETE",
@@ -68,7 +82,7 @@ export default function EditCard(props) {
             history.push("/error")
         } else {
             updateProjectSocket()
-            props.hideFormEdit()
+            props.hideForm()
         }
 
     }, [history, props, cardId, listId, updateProjectSocket])
@@ -161,6 +175,11 @@ export default function EditCard(props) {
                             <span className={styles.pic2}>
                                 <img src={pic2} alt="pic2" />
                             </span>
+                            <span className={styles.nameContainer}>
+                                <p  >Progress</p>
+                            </span>
+                        </div>
+                        <div>
                             {
                                 isProgressActive ?
                                     <div ref={dropdownRef}>
@@ -169,26 +188,23 @@ export default function EditCard(props) {
                                             <button onClick={handleSubmit} className={styles.editButton} >Edit</button>
                                         </span></div>
                                     :
-                                    <span className={styles.nameContainer}>
-                                        <p className={styles.textName} onClick={() => setIsProgressActive(true)} >Progress</p>
-                                    </span>
-                            }
-                        </div>
+                                    <div className={styles.progressDiv} onClick={() => setIsProgressActive(true)}>
+                                        {
+                                            card.progress ?
+                                                <div className={styles.bar} >
+                                                    <div
+                                                        style={{
+                                                            width: `${card.progress}%`,
+                                                            ['backgroundColor']: progressColor(card.progress)
+                                                        }}
+                                                        className={styles.progress}
+                                                    />
+                                                </div> : null
+                                        }
+                                        <span className={styles.textName} >{card.progress} %</span>
+                                    </div>
 
-                        <div className={styles.progressDiv}>
-                            {
-                                card.progress ?
-                                    <div className={styles.bar} >
-                                        <div
-                                            style={{
-                                                width: `${card.progress}%`,
-                                                ['backgroundColor']: progressColor(card.progress)
-                                            }}
-                                            className={styles.progress}
-                                        />
-                                    </div> : null
                             }
-                            <span>{card.progress} %</span>
                         </div>
 
                     </div>
@@ -227,16 +243,16 @@ export default function EditCard(props) {
 
 
                     <div className={styles.lasRow}>
-                        <Button onClick={handleSubmit} title="Edit Task" />
-                        <Button onClick={(e) => { if (window.confirm('Are you sure you wish to delete this item?')) deleteCard(e) }} title="Delete Task" />
-                        <Button onClick={cancelSubmit} title="Cancel" />
+
                     </div>
 
                 </div>
 
                 <div className={styles.rightSide}>
-                    <div className={styles.membersContainer}>
-                        <div className={styles.inputTitles}>
+
+
+                    <div className={styles.membersDiv}>
+                        <div >
                             <p className={styles.text}>Members</p>
                         </div>
                         <TaskMembers cardMembers={members} size={30} cardId={cardId} listId={listId} project={props.project} title={'Add'} />
@@ -244,14 +260,58 @@ export default function EditCard(props) {
 
 
                     <div className={styles.secondRow}>
-                        <div className={styles.inputTitles}>
+                        <div >
                             <p className={styles.text}>Due Date</p>
+                            <div className={styles.dueDate} >
+                                <TaskDueDate cardDueDate={dueDate} cardId={cardId} listId={listId} project={props.project} />
+                            </div>
+
                         </div>
-                        <TaskDueDate cardDueDate={dueDate} cardId={cardId} listId={listId} />
                     </div>
 
-                </div>
 
+                    <div className={styles.thirdRow}>
+                        <div>
+                            <p className={styles.text}>Add</p>
+                        </div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic3} alt="pic3" />
+                            Join</div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic4} alt="pic4" />
+                            Stickers</div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic5} alt="pic5" />
+                            Due Date</div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic6} alt="pic6" />
+                            Attach file</div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic7} alt="pic7" />
+                            Reports</div>
+                    </div>
+
+
+                    <div className={styles.lasRowRight}>
+                        <div>
+                            <p className={styles.text}>Manage</p>
+                        </div>
+                        <div className={styles.smallButtonsNoPic} >
+                            Make Template</div>
+                        <div className={styles.smallButtonsNoPic} >
+                            Remove Note</div>
+                        <button className={styles.smallButtonsNoPic} onClick={(e) => { deleteCard(e) }} title="Delete Task" >
+                            Delete Task</button>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic8} alt="pic8" />
+                            Settings</div>
+                        <div className={styles.smallButtons} >
+                            <img className={styles.picsSmallButtons} src={pic9} alt="pic9" />
+                            View</div>
+                    </div>
+
+
+                </div>
 
             </form>
         </div>
